+++ src/common/input.ts
@@ -1321,7 +1321,7 @@ export class Input extends EventEmitter {
         });
 
         this.client.on('add-line', (data) => {
-            this.ExecuteTriggers(TriggerTypes.Regular | TriggerTypes.Pattern, data.line, data.raw, data.fragment, false, true);
+            this.ExecuteTriggers(TriggerTypes.Regular | TriggerTypes.Pattern | TriggerTypes.LoopExpression, data.line, data.raw, data.fragment, false, true);
             if (this._gag > 0 && !data.fragment) {
                 data.gagged = true;
                 this._gag--;
@@ -1858,7 +1858,8 @@ export class Input extends EventEmitter {
                 if ((<any>this.client).sendChat)
                     (<any>this.client).sendChat(args);
                 return null;
-            //spell-checker:ignore untrigger
+            //spell-checker:ignore untrigger unaction
+            case 'unaction':
             case 'untrigger':
             case 'unt':
                 profile = null;
@@ -2045,6 +2046,7 @@ export class Input extends EventEmitter {
                                     case 'alarm':
                                     case 'event':
                                     case 'cmdpattern':
+                                    case 'loopexpression':
                                         //case 'expression':
                                         item.options[o.trim()] = true;
                                         break;
@@ -2100,6 +2102,7 @@ export class Input extends EventEmitter {
                                     case 'alarm':
                                     case 'event':
                                     case 'cmdpattern':
+                                    case 'loopexpression':
                                         //case 'expression':
                                         item.options[o.trim()] = true;
                                         break;
@@ -4416,7 +4419,7 @@ export class Input extends EventEmitter {
                 return null;
             case 'fire':
                 args = this.parseInline(args.join(' ') + '\n');
-                this.ExecuteTriggers(TriggerTypes.Regular | TriggerTypes.Pattern, args, args, false, false);
+                this.ExecuteTriggers(TriggerTypes.Regular | TriggerTypes.Pattern | TriggerTypes.LoopExpression, args, args, false, false);
                 return null;
             case 'state': //#STATE id state profile
             case 'sta':
@@ -4830,6 +4833,7 @@ export class Input extends EventEmitter {
                                     case 'alarm':
                                     case 'event':
                                     case 'cmdpattern':
+                                    case 'loopexpression':
                                     //case 'expression':
                                     case 'reparse':
                                     case 'reparsepattern':
@@ -4894,6 +4898,7 @@ export class Input extends EventEmitter {
                                     case 'alarm':
                                     case 'event':
                                     case 'cmdpattern':
+                                    case 'loopexpression':
                                     //case 'expression':
                                     case 'reparse':
                                     case 'reparsepattern':
@@ -7518,6 +7523,8 @@ export class Input extends EventEmitter {
             return true;
         if (type === TriggerType.Regular && (types & TriggerTypes.Regular) == TriggerTypes.Regular)
             return true;
+        if (type === TriggerType.LoopExpression && (types & TriggerTypes.LoopExpression) == TriggerTypes.LoopExpression)
+            return true;
         //if (type === TriggerType.Expression && (types & TriggerTypes.Expression) == TriggerTypes.Expression)
         //return true;            
         return false;
@@ -7528,8 +7535,6 @@ export class Input extends EventEmitter {
             return true;
         if ((type & SubTriggerTypes.Wait) == SubTriggerTypes.Wait)
             return true;
-        //if ((type & SubTriggerTypes.LoopExpression) == SubTriggerTypes.LoopExpression)
-        //return true;
         if ((type & SubTriggerTypes.LoopPattern) == SubTriggerTypes.LoopPattern)
             return true;
         if ((type & SubTriggerTypes.LoopLines) == SubTriggerTypes.LoopLines)
@@ -7667,9 +7672,38 @@ export class Input extends EventEmitter {
                         continue;
                     }
                 }
+                /*
+                else if (this._TriggerStates[t].type === TriggerType.LoopExpression) {
+                    //move on after line count
+                    if (this._TriggerStates[t].loop != -1 && this._TriggerStates[t].lineCount < 1) {
+                        this.advanceTrigger(trigger, parent, t);
+                        //reparse as new state may be valid
+                        if (!this._TriggerStates[t])
+                            this._TriggerStates[t] = { reParse: true }
+                        else
+                            this._TriggerStates[t].reParse = true;
+                        t = this.cleanUpTriggerState(t);
+                        continue;
+                    }
+                }
+                */
             }
             try {
-                if (trigger.verbatim) {
+                if (trigger.type === TriggerType.LoopExpression) {
+                    if (this.evaluate(trigger.pattern)) {
+                        if (!this._TriggerStates[t])
+                            this._TriggerStates[t] = this.createTriggerState(trigger, false, parent);
+                        else if (this._TriggerStates[t].loop !== -1 && this._TriggerStates[t].lineCount < 1)
+                            continue;
+                        val = this.ExecuteTrigger(trigger, [(trigger.raw ? raw : line)], ret, t, [(trigger.raw ? raw : line)], 0, parent);
+                    }
+                    else {
+                        //this.updateTriggerState(trigger, t);
+                        this.advanceTrigger(trigger, parent, t);
+                        continue;
+                    }
+                }
+                else if (trigger.verbatim) {
                     if (!trigger.caseSensitive && (trigger.raw ? raw : line).toLowerCase() !== trigger.pattern.toLowerCase()) {
                         //if reparse and if failed advance anyways
                         if (!this._TriggerStates[t] && (trigger.type === SubTriggerTypes.ReParse || trigger.type === SubTriggerTypes.ReParsePattern)) {
@@ -7959,6 +7993,14 @@ export class Input extends EventEmitter {
             }
             else if (this._TriggerStates[idx].type === SubTriggerTypes.WithinLines)
                 this.clearTriggerState(idx);
+            else if (this._TriggerStates[idx].type === TriggerType.LoopExpression) {
+                //infintate until expression is false
+                if (this._TriggerStates[idx].loop === -1)
+                    return;
+                //else if uses a line count
+                if (this._TriggerStates[idx].lineCount > 0)
+                    return;
+            }
         }
         parent.state++;
         //1 based
@@ -7972,7 +8014,7 @@ export class Input extends EventEmitter {
             this._TriggerStates[idx] = this.createTriggerState(parent.triggers[parent.state - 1]);
     }
 
-    public createTriggerState(trigger, reparse?) {
+    public createTriggerState(trigger, reparse?, parent?) {
         let params;
         let state;
         switch (trigger.type) {
@@ -8017,7 +8059,6 @@ export class Input extends EventEmitter {
                 state = { remoteCount: params + 1 };
                 break;
             */
-            //case SubTriggerTypes.LoopExpression:
             case SubTriggerTypes.LoopPattern:
                 params = trigger.params;
                 if (params && params.length) {
@@ -8028,6 +8069,20 @@ export class Input extends EventEmitter {
                 else
                     params = 0;
                 state = { loop: params };
+                break;
+            case TriggerType.LoopExpression:
+                params = trigger.params;
+                if (params && params.length) {
+                    params = parseInt(params, 10);
+                    if (isNaN(params))
+                        params = 1;
+                    if(parent === trigger)
+                        state = { lineCount: params - 1};
+                    else
+                        state = { lineCount: params };
+                }
+                else
+                    state = { loop: -1 };
                 break;
         }
         if (state)
@@ -8080,7 +8135,6 @@ export class Input extends EventEmitter {
                             this._TriggerStates[idx].remoteCount = params;
                             break;
             */
-            //case SubTriggerTypes.LoopExpression:
             case SubTriggerTypes.LoopPattern:
                 params = trigger.params;
                 if (params && params.length) {
@@ -8091,6 +8145,17 @@ export class Input extends EventEmitter {
                 else
                     params = 0;
                 this._TriggerStates[idx].loop = params;
+                break;
+            case TriggerType.LoopExpression:
+                params = trigger.params;
+                if (params && params.length) {
+                    params = parseInt(params, 10);
+                    if (isNaN(params))
+                        params = 1;
+                    this._TriggerStates[idx].lineCount = params + 1;
+                }
+                else
+                    this._TriggerStates[idx].loop = -1;
                 break;
         }
     }
@@ -8446,8 +8511,10 @@ export class Input extends EventEmitter {
                     sTrigger.type = TriggerType.Event;
                 if (options.cmdpattern)
                     sTrigger.type = TriggerType.CommandInputPattern;
+                if (options.loopexpression)
+                    sTrigger.type = TriggerType.LoopExpression;
                 //if(options.expression)
-                //trigger.type = TriggerType.Expression;                    
+                //sTrigger.type = TriggerType.Expression;                    
                 if (options.reparse)
                     sTrigger.type = SubTriggerTypes.ReParse;
                 if (options.reparsepattern)
@@ -8525,6 +8592,8 @@ export class Input extends EventEmitter {
                     trigger.type = TriggerType.Event;
                 if (options.cmdpattern)
                     trigger.type = TriggerType.CommandInputPattern;
+                if (options.loopexpression)
+                    sTrigger.type = TriggerType.LoopExpression;
                 //if(options.expression)
                 //trigger.type = TriggerType.Expression;